CAMEL-17306: camel-salesforce - Refactor DTO code generation
