+++ components/camel-salesforce/camel-salesforce-codegen/src/main/resources/sobject-query-records.vm
@@ -1,0 +1,44 @@
+## ------------------------------------------------------------------------
+## Licensed to the Apache Software Foundation (ASF) under one or more
+## contributor license agreements.  See the NOTICE file distributed with
+## this work for additional information regarding copyright ownership.
+## The ASF licenses this file to You under the Apache License, Version 2.0
+## (the "License"); you may not use this file except in compliance with
+## the License.  You may obtain a copy of the License at
+##
+## http://www.apache.org/licenses/LICENSE-2.0
+##
+## Unless required by applicable law or agreed to in writing, software
+## distributed under the License is distributed on an "AS IS" BASIS,
+## WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+## See the License for the specific language governing permissions and
+## limitations under the License.
+## ------------------------------------------------------------------------
+## sobject-query-records.vm
+/*
+ * Salesforce Query DTO generated by camel-salesforce-maven-plugin
+ */
+package $packageName;
+
+import org.apache.camel.component.salesforce.api.dto.AbstractQueryRecordsBase;
+
+import java.util.List;
+import javax.annotation.Generated;
+
+/**
+ * Salesforce QueryRecords DTO for type $desc.Name
+ */
+#set( $descName = $desc.Name )
+@Generated("org.apache.camel.maven.CamelSalesforceMojo")
+public class QueryRecords$descName extends AbstractQueryRecordsBase {
+
+    private List<$descName> records;
+
+    public List<$descName> getRecords() {
+        return records;
+    }
+
+    public void setRecords(List<$descName> records) {
+        this.records = records;
+    }
+}
