+++ components/camel-salesforce/pom.xml
@@ -34,6 +34,7 @@
 
     <modules>
         <module>camel-salesforce-component</module>
+        <module>camel-salesforce-codegen</module>
         <module>camel-salesforce-maven-plugin</module>
     </modules>
