+++ parent/pom.xml
@@ -2261,6 +2261,11 @@
 			</dependency>
 			<dependency>
 				<groupId>org.apache.camel</groupId>
+				<artifactId>camel-salesforce-codegen</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>org.apache.camel</groupId>
 				<artifactId>camel-salesforce-maven-plugin</artifactId>
 				<version>${project.version}</version>
 			</dependency>
